Use int instead of short for index

+++ limber-backend-application/module/forms/api/src/main/kotlin/io/limberapp/backend/module/forms/service/formTemplate/FormTemplatePartService.kt
@@ -5,7 +5,7 @@ import java.util.UUID
 
 interface FormTemplatePartService {
 
-    fun create(formTemplateId: UUID, model: FormTemplatePartModel, index: Short? = null)
+    fun create(formTemplateId: UUID, model: FormTemplatePartModel, index: Int? = null)
 
     fun update(
         formTemplateId: UUID,
