+++ piper/exceptions/src/main/kotlin/com/piperframework/exception/exception/badRequest/IndexOutOfBounds.kt
@@ -1,3 +1,3 @@
 package com.piperframework.exception.exception.badRequest
 
-class IndexOutOfBounds(index: Short) : BadRequestException("The given index \"$index\" is out of bounds.")
+class IndexOutOfBounds(index: Int) : BadRequestException("The given index \"$index\" is out of bounds.")
